trigger_functions/requirements.txt
firebase-admin==4.4.0
google-api-python-client==1.8.0
google-auth==1.25
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-tasks==1.5.0
google-python-cloud-debugger; sys_platform == 'linux'
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.3
mypy-extensions==0.4.3
psycopg2==2.8.4
python-dotenv==0.12.0
pytz==2020.4
requests==2.23.0
SQLAlchemy==1.3.23
Werkzeug==0.16.1
python-magic==0.4.18
python-dateutil==2.8.2
